+++ network/session_test.go
@@ -6,21 +6,21 @@ import (
 
 func Test_OptSession(t *testing.T) {
 	NewSessionMap()
-	ret := OptSession(uint32(1), Find, []SessionData{SessionData{
+	ret := OptSession(uint64(1), Find, []SessionData{SessionData{
 		Key: "key",
 	}})
 	if ret != nil {
 		t.Error("user 1 should be nil")
 	}
 	// add user 1
-	OptSession(uint32(1), Update, nil)
+	OptSession(uint64(1), Update, nil)
 	// add user 1
-	OptSession(uint32(1), Update, []SessionData{SessionData{
+	OptSession(uint64(1), Update, []SessionData{SessionData{
 		Key:   "key1",
 		Value: 9257,
 	}})
 	// add user 1
-	OptSession(uint32(1), Update, []SessionData{SessionData{
+	OptSession(uint64(1), Update, []SessionData{SessionData{
 		Key:   "key2",
 		Value: "hello",
 	}, SessionData{
@@ -28,17 +28,17 @@ func Test_OptSession(t *testing.T) {
 		Value: 128.123123,
 	}})
 	// find
-	ret1 := OptSession(uint32(1), Find, nil)
+	ret1 := OptSession(uint64(1), Find, nil)
 	if ret1 != nil {
 		t.Error("find error")
 	}
-	ret2 := OptSession(uint32(1), Find, []SessionData{SessionData{
+	ret2 := OptSession(uint64(1), Find, []SessionData{SessionData{
 		Key: "key2",
 	}})
 	if ret2 == nil || len(ret2) != 1 || ret2[0].Value != "hello" {
 		t.Error("find error")
 	}
-	ret3 := OptSession(uint32(1), Find, []SessionData{SessionData{
+	ret3 := OptSession(uint64(1), Find, []SessionData{SessionData{
 		Key: "key2",
 	}, SessionData{
 		Key: "key3",
@@ -52,14 +52,14 @@ func Test_OptSession(t *testing.T) {
 		t.Error("find error")
 	}
 	// update key3 key4
-	OptSession(uint32(1), Update, []SessionData{SessionData{
+	OptSession(uint64(1), Update, []SessionData{SessionData{
 		Key:   "key3",
 		Value: 12222, // 128.123123
 	}, SessionData{
 		Key:   "key4",
 		Value: "xixixixiixiix",
 	}})
-	ret4 := OptSession(uint32(1), Find, []SessionData{SessionData{
+	ret4 := OptSession(uint64(1), Find, []SessionData{SessionData{
 		Key: "key3",
 	}, SessionData{
 		Key: "key4",
@@ -71,10 +71,10 @@ func Test_OptSession(t *testing.T) {
 		t.Error("find error")
 	}
 	// delete ke3
-	OptSession(uint32(1), Delete, []SessionData{SessionData{
+	OptSession(uint64(1), Delete, []SessionData{SessionData{
 		Key: "key3",
 	}})
-	ret5 := OptSession(uint32(1), Find, []SessionData{SessionData{
+	ret5 := OptSession(uint64(1), Find, []SessionData{SessionData{
 		Key: "key3",
 	}, SessionData{
 		Key: "key4",
